+++ R/min_possible_prevalence.R
@@ -23,41 +23,33 @@
 #' you need a prevalence of at least 1 out of 21"
 min_possible_prevalence <- function(Sensitivity, FP_test, min_PPV_desired) {
 
-  #TEST Parameters **************
-    #FP
-    Max_FP = 100
-    Steps_FP = 1000
-    Step_size_FP = Max_FP / Steps_FP
-    Min_FP = 0
-    FP = seq(Min_FP, Max_FP, Step_size_FP)
 
-    #CONDITION Parameters ***********
+  # Fixed parameters --------------------------------------------------------
 
-    #Prevalence_y - x out of y
-    Prevalence_x = 1
     Min_Prevalence = 1
     Max_Prevalence = 10000 # CHANGE ME
     Steps_Prevalence = 10000
     Step_size_Prevalence = Max_Prevalence / Steps_Prevalence
     Prevalence = seq(Min_Prevalence, (1 + Max_Prevalence), Step_size_Prevalence)
 
-    # ****************************************************************************************
+
+  # Calculation -------------------------------------------------------------
 
     # We calculate the 100x100 PPV matrix
-    PPV = (Sensitivity * Prevalence_x) / ( (Sensitivity * Prevalence_x) + ( (Prevalence - 1) %o% FP) )
-    #Label columns and rows of matrix
-    colnames(PPV) = FP
-    rownames(PPV) = Prevalence
+    PPV = (Sensitivity * Min_Prevalence) / ( (Sensitivity * Min_Prevalence) + ( (Prevalence - 1) * FP_test) )
 
-    # Long format para ggplot Heatmap
-    PPV_melted = reshape2::melt(PPV)
+    # Long format 
+    PPV_melted = PPV %>% as.data.frame() %>% mutate(Prevalence = 1:length(.))
     
     # Rename columns
-    names(PPV_melted) = c("melted_Prevalence", "melted_FP", "melted_PPV")
+    names(PPV_melted) = c("melted_PPV", "melted_Prevalence")
 
     # Calculate prevalence
-    output_prevalence = max(PPV_melted$melted_Prevalence[PPV_melted$melted_PPV > (min_PPV_desired / 100) & PPV_melted$melted_FP == FP_test])
+    output_prevalence = max(PPV_melted$melted_Prevalence[PPV_melted$melted_PPV > (min_PPV_desired / 100)])
+    # PPV_melted %>% filter(abs(melted_PPV - (min_PPV_desired / 100)) == min(abs(melted_PPV - (min_PPV_desired / 100)))) # Keep closest value to min_PPV_desired
 
-    # Function output!
+    
+  # Function output --------------------------------------------------------
     message("To reach a PPV of ", min_PPV_desired, "% when using a test with ", Sensitivity, "% Sensitivity and ", FP_test, "% False Positive Rate, you need a prevalence of at least 1 out of ", output_prevalence)
+    
 }